+++ app/user/recruitments/(my)/layout.tsx
@@ -1,0 +1,16 @@
+import type { ReactNode } from "react";
+
+import styles from "./layout.module.scss";
+
+import Sidebar from "../_components/sidebar/sidebar";
+
+export default function MypageLayout({ children }: { children: ReactNode }) {
+  return (
+    <div className={styles.layout}>
+      <section className={styles.layout__sidebar}>
+        <Sidebar />
+      </section>
+      <main className={styles.layout__content}>{children}</main>
+    </div>
+  );
+}
